Fix IsTrigger Bug
Bug -
If the hero was unable to move between borders at a specific time, it would disable the border completely (preventing future use of the border, regardless of the adjacent frame).

Background for the bug -
In order to prevent the hero from going through the borders (instead of moving between frames), the isTrigger property had to be set as "false" once the trigger was activated.
the isTrigger property was set to be enabled again once the OnExit trigger occured, however, in some cases it didn't work, which lef the property as "false".

Fix -
Duplicated the borders:
From the original one, removed the script and triggers.
From the new one, removed the mesh (so it's invisible).

Removed the isTrigger changes in the script.