Changed frame movement controls to arrows AND moved hero when moving a frame
Change frame movement controls to arrows:
added a zoom in\out check in the hero's movement script:
when zoomed in - move the hero.
when zoomed out - disabled hero movement

move hero with moved frames:
added an "active frame" field to the Frames manager which keeps up with whichever frame the hero is currently in.
Thus, it is possible to get the hero's relative location to it's current frame and keep it while moving the frame