Virgin movement - escape from hero is implemented
This flow happens when the hero enters another frame:
1. the system checks if the hero is in the same frame as the virgin. if not, the flow is done.
(the next steps are availalbe only if they are in the same frame)
2. the system checks for possible adjacent frames that the virgin can move to by their border labels (note - the virgin cannot move to the frame the hero just came from)
if there are no possible frames, the win/lose functions are triggered (not yet implemented).
3. the virgin will attempt to move as far away from the hero as possible.
for example: if the hero came from the LEFT frame, the virgin will move to the RIGHT frame (if able).
if the furthest frame isn't possible, the virgin will go to the first possible frame from the following order LEFT > TOP > RIGHT > BOTTOM (excluding the frame the hero came from).

the virgin will be in the same location INSIDE the frame as in the previous one (in case we want her to be on a platform for example) - this is relevant to level design, as we don't want her to fall into the abyss or get stuck in a wall for some reason.