Added Puzzle Manager and Small Puzzle Logic
Each level has several small puzzles (0 or more).

each small puzzle can either be dependent or independent.
if a small puzzle is dependent, then all previous puzzles need to be solved before the dependent puzzle can be solved.

the dependency is set in the inspector (in the puzzle manager).
each small puzzle has an indicator for it's place in the list, which is also configured in the inspector (for each puzzle)

whenever the hero triggers a small puzzle (will be defined later) the following happens:
1. the system checks if the small puzzle is already solved
(if it is, then it stops here).

2. if the small puzzle is not solved, the system checks if the small puzzle can be solved:
A small puzzle can be solved in the following cases:
a. if it is not dependent (or if it is the first puzzle in the list)
b. if it is dependent and all previous small puzzles are already solved

3. if the puzzle can be solved, then it is marked as solved in the list

4. last, the system checks if all small puzzles are solved in the level, if they are, then the levelPuzzle is marked as solved (the levelPuzzle is checked when the hero enters the puzzle room with the virgin, to indicate if the player has won the level)