Added Game Over spawn mechanic, OppositeDirectionDict, UIManager
Game Over mechanic - once "Try Again" button is clicked, the hero spawns in the previous frame. This is done by saving the previous position each time the hero moves. The intial value of previousHeroFrame (or after "Try Again") is (-1,-1)

oppositeDirectionDict - a dictionary FramesManager which gives each direction of entry, its opposite direction. I used this dictionary only in checkLabels(), there might be other useful uses as well.

UIManager - manager for panels. Can be merged with Imagemanager in the future.

BUG - Sometimes SwitchHeroFrame is triggered twice  when hero enters the puzzle room and loses. This causes the hero to spawn at the same frame instead of the previous one (because it makes the previous frame = current frame)